Challenges/Day17/docker-flask-app/requirements.txt
# Core Flask & Web Server
flask==2.3.2
gunicorn==20.1.0

# Database
psycopg2-binary==2.9.6
flask-sqlalchemy==3.0.3
alembic==1.11.1  # For database migrations

# Environment & Configuration
python-dotenv==1.0.0

# Monitoring & Metrics
prometheus-client==0.16.0
blinker==1.6.2  # For Flask signals

# Security
bcrypt==4.0.1  # For password hashing
python-jose==3.3.0  # JWT support

# Utilities
python-dateutil==2.8.2  # Date parsing
requests==2.31.0  # HTTP client
click==8.1.3  # CLI support

# Development (optional)
pytest==7.4.0
pytest-cov==4.1.0
faker==18.12.0  # Test data generation